backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "cryptography>=46.0.5",
    "email-validator>=2.2.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "openai>=2.20.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.33",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.40.0",
]
